refact loop
- MainView -> CustomerView, EmployeeView
- AppConfig View에 대해서 의존성 주입 로직 추가
