+++ src/loop/AppConfig.java
@@ -2,18 +2,31 @@ package loop;
 
 import customer.CustomerControllerImpl;
 import customer.CustomerRepository;
+import employee.EmployeeControllerImpl;
 import employee.EmployeeRepository;
+import parcel.ParcelRepository;
 
 public class AppConfig {
 	public View view() {
-		return new MainViewImpl(customerController(), employeeController());
+		return new MainViewImpl(customerView(), employeeView());
 	}
 
-	private EmployeeController employeeController() {
-		return new EmployeeController(employeeRepository());
+	private EmployeeVewImpl employeeView() {
+		return new EmployeeVewImpl(employeeController(), parcelController());
+	}
+
+	private CustomerViewImpl customerView() {
+		return new CustomerViewImpl(customerController(), parcelController());
+	}
+
+	private EmployeeControllerImpl employeeController() {
+		return new EmployeeControllerImpl(employeeRepository());
 	}
 	private CustomerControllerImpl customerController() {
 		return new CustomerControllerImpl(customerRepository());
+	}
+	private ParcelControllerImpl parcelController() {
+		return new ParcelControllerImpl(parcelRepository());
 	}
 	private CustomerRepository customerRepository() {
 		return new CustomerRepository();
@@ -21,4 +34,8 @@ public class AppConfig {
 	private EmployeeRepository employeeRepository() {
 		return new EmployeeRepository();
 	}
+	private ParcelRepository parcelRepository() {
+		return new ParcelRepository();
+	}
+
 }
