+++ src/loop/Repository.java
@@ -4,6 +4,6 @@ package loop;
  * packageName    : loop fileName       : CustomerRepository author         : hoho date           :
  * 4/26/24 description    :
  */
-public interface Repository {
+public interface  Repository {
 
 }
